infra-portal-be/requirements.txt
fastapi
gunicorn
uvicorn-worker
uvicorn
motor
pydantic
pymongo
annotated-types
python-dotenv